summarization_cli/requirements.txt
openai
python-dotenv
azure-cosmos